Public snapshot: sync with dev master (2026-06-11)
Sprints 365-384: capability ladder Gates A-F (compiler, console I/O,
arrays, Brainfuck), perf track P1-P4 + dirty-residue plumbing (~81x),
active-interconnect synthesis S1-S5, AlphaFabric placement env + HLS
middle layer, SIMT Tier 0 baseline, QASM3 parser hardening.

+++ crates/v2-viz/src/sim.rs
@@ -18,7 +18,7 @@ pub struct SimState {
 #[derive(Resource)]
 pub struct SimSnapshot {
     pub cycle: u64,
-    pub pc: u8,
+    pub pc: u32,
     pub flag_z: bool,
     pub flag_c: bool,
     pub halted: bool,
